Use eval to create autocmd's for the K mappings.
Use the eval functionality in vim to dynamically create the auto
commands for the K mappings for each file type specified in the
g:dokumentary_docprgs variable.

This allows the user to properly customise which command to use for each
file type, and even add more file types by adding more keys to
g:dokumentary_docprgs.

Closes #2.